+++ back/src/main/java/com/udteam/miristock/entity/ArticleEntity.java
@@ -18,7 +18,7 @@ public class ArticleEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long articleNo; // PK
+    private Integer articleNo; // PK
 
     @Column(columnDefinition = "TEXT", nullable = false)
     private String memberNickname; // 회원 닉네임
